refactor: update bigdecimal for id to long

+++ src/main/java/com/htdinh/bookstore/dto/request/LogoutRequest.java
@@ -3,10 +3,9 @@ package com.htdinh.bookstore.dto.request;
 import lombok.Getter;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 @Getter
 public class LogoutRequest {
     @NotNull
-    private BigDecimal id;
+    private Long id;
 }
